fixed plot/legend dimensions, added sliders to modify fig size
All of the elements in the plot window now fit together nicely without
any overlap, going out of bounds, or excess blank space.  This is
generalized to work with any plot size, which can now be played with via
sliders below the plot.  Additional small changes were made to beautify
things.  A button to save figures locally as PDFs was also added.